feat(s3_cuobject): isolate RDMA into a dedicated provider subclass
Moves the entire S3-over-RDMA (cuObject) data plane out of the s3 provider
into a new s3_cuobject provider that subclasses S3StorageProvider and overrides
only object GET/PUT. The base s3 provider is now byte-identical to main again,
so the RDMA path -- which cannot be exercised in CI without an RDMA NIC + an
RDMA-capable endpoint -- can never affect it. Mirrors the s8k/ais_s3/gcs_s3
subclass pattern. Enable with `type: s3_cuobject` instead of `type: s3` +
`rdma: {}`. Registered in config/schema/providers; RDMA unit tests migrated.
