+++ multi-storage-client/src/multistorageclient/providers/s3.py
@@ -13,11 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import base64
 import codecs
 import io
 import os
-import struct
 import tempfile
 from collections.abc import Callable, Iterator
 from typing import IO, Any, Optional, TypeVar, Union
@@ -52,7 +50,6 @@ from ..utils import (
     split_path,
     validate_attributes,
 )
-from ._cuobj import CuObjEngine
 from .base import BaseStorageProvider
 
 _T = TypeVar("_T")
@@ -65,17 +62,6 @@ MULTIPART_CHUNKSIZE = 32 * MiB
 IO_CHUNKSIZE = 32 * MiB
 PYTHON_MAX_CONCURRENCY = 8
 MAX_POOL_CONNECTIONS = DEFAULT_MAX_POOL_CONNECTIONS
-
-# When RDMA is enabled, small objects and every GET move as a single registered
-# buffer, so the boto multipart threshold is raised past any practical object
-# size and those transfers take the single-shot path. Larger uploads are split
-# by the RDMA multipart path below (single PutObject caps at 5 GiB, and parts
-# bound the size of any one registered buffer / RDMA transfer).
-RDMA_SINGLE_SHOT_THRESHOLD = 1 << 62
-
-# Default RDMA multipart part size. Uploads larger than this are sent as an
-# RDMA multipart upload (one registered buffer + token + CRC64NVME per part).
-RDMA_MULTIPART_CHUNKSIZE = 512 * MiB
 
 PROVIDER = "s3"
 
@@ -225,32 +211,14 @@ class S3StorageProvider(BaseStorageProvider):
         self._verify = verify
 
         self._signature_version = kwargs.get("signature_version", "s3v4")
-
-        # S3-over-RDMA (NVIDIA cuObject) data plane. Mutually exclusive with the
-        # Rust client; forces the empty-body, unsigned-payload, no-checksum wire
-        # contract the RDMA endpoint expects.
-        self._rdma_options = kwargs.get("rdma")
-        if self._rdma_options is not None and "rust_client" in kwargs:
-            raise ValueError("The 's3' provider options 'rdma' and 'rust_client' are mutually exclusive.")
-        request_checksum_calculation = kwargs.get("request_checksum_calculation")
-        response_checksum_validation = kwargs.get("response_checksum_validation")
-        s3_config = kwargs.get("s3")
-        if self._rdma_options is not None:
-            overrides = CuObjEngine.client_config_overrides()
-            request_checksum_calculation = overrides["request_checksum_calculation"]
-            response_checksum_validation = overrides["response_checksum_validation"]
-            # Only the empty-body wire contract (payload_signing_enabled=False) is
-            # enforced; addressing style stays user-controlled via the `s3` option.
-            s3_config = {**(s3_config or {}), **overrides["s3"]}
-
         self._s3_client = self._create_s3_client(
-            request_checksum_calculation=request_checksum_calculation,
-            response_checksum_validation=response_checksum_validation,
+            request_checksum_calculation=kwargs.get("request_checksum_calculation"),
+            response_checksum_validation=kwargs.get("response_checksum_validation"),
             max_pool_connections=kwargs.get("max_pool_connections", MAX_POOL_CONNECTIONS),
             connect_timeout=kwargs.get("connect_timeout", DEFAULT_CONNECT_TIMEOUT),
             read_timeout=kwargs.get("read_timeout", DEFAULT_READ_TIMEOUT),
             retries=kwargs.get("retries"),
-            s3=s3_config,
+            s3=kwargs.get("s3"),
         )
         self._transfer_config = TransferConfig(
             multipart_threshold=int(kwargs.get("multipart_threshold", MULTIPART_THRESHOLD)),
@@ -264,21 +232,6 @@ class S3StorageProvider(BaseStorageProvider):
         self._signer_cache: dict[tuple, URLSigner] = {}
 
         self._checksum_algorithm: Optional[str] = self._validate_checksum_algorithm(kwargs.get("checksum_algorithm"))
-
-        self._rdma_engine: Optional[CuObjEngine] = None
-        self._rdma_multipart_chunksize = RDMA_MULTIPART_CHUNKSIZE
-        if self._rdma_options is not None:
-            self._rdma_engine = CuObjEngine()
-            self._rdma_engine.install_hooks(self._s3_client)
-            self._checksum_algorithm = None
-            self._multipart_threshold = RDMA_SINGLE_SHOT_THRESHOLD
-            self._rdma_multipart_chunksize = int(
-                self._rdma_options.get("multipart_chunksize", RDMA_MULTIPART_CHUNKSIZE)
-            )
-            if self._rdma_multipart_chunksize < 1:
-                raise ValueError(
-                    f"rdma.multipart_chunksize must be a positive integer, got {self._rdma_multipart_chunksize}"
-                )
 
         self._rust_client = None
         if "rust_client" in kwargs:
@@ -547,174 +500,6 @@ class S3StorageProvider(BaseStorageProvider):
                 f"Failed to {operation} object(s) at {bucket}/{key}, error type: {type(error).__name__}, error: {error}"
             ) from error
 
-    @staticmethod
-    def _rdma_checksum(buffer) -> str:
-        """Base64 CRC64NVME of ``buffer`` for the ``x-amz-checksum-crc64nvme`` header.
-
-        CRC64NVME is hardware-accelerated and computed via ``awscrt`` (the same
-        implementation botocore uses for this algorithm).
-        """
-        try:
-            from awscrt import checksums
-        except ImportError as error:
-            raise RuntimeError(
-                "RDMA PUT computes a CRC64NVME checksum and requires the 'awscrt' package (pip install awscrt)."
-            ) from error
-        return base64.b64encode(struct.pack(">Q", checksums.crc64nvme(buffer))).decode("ascii")
-
-    def _rdma_put(self, kwargs: dict[str, Any], body: bytes) -> int:
-        """Single-shot RDMA PUT: cuObject transfers the registered buffer; the HTTP body is empty.
-
-        A CRC64NVME checksum of the payload is computed on the client and sent as
-        ``x-amz-checksum-crc64nvme``. On an RDMA-capable endpoint it is validated
-        against the bytes delivered over RDMA; on an endpoint that ignores the
-        RDMA token it is validated against the (empty) HTTP body and fails with a
-        checksum mismatch, so the upload is rejected rather than silently storing
-        a 0-byte object.
-        """
-        engine = self._rdma_engine
-        assert engine is not None
-        # cuObject pins a writable region: reuse the caller's buffer when it is
-        # writable, and copy only a read-only (immutable) body such as bytes.
-        buffer = bytearray(body) if memoryview(body).readonly else body
-        if len(buffer) == 0:
-            self._s3_client.put_object(**kwargs)
-            return 0
-        kwargs = {**kwargs, "Body": b"", "ChecksumCRC64NVME": self._rdma_checksum(buffer)}
-        with engine.transfer(buffer, is_put=True):
-            response = self._s3_client.put_object(**kwargs)
-            engine.check_reply(response)
-        return len(buffer)
-
-    def _rdma_get(self, path: str, bucket: str, key: str, byte_range: Optional[Range]) -> bytearray:
-        """Single-shot RDMA GET into a registered buffer; returns the buffer."""
-        engine = self._rdma_engine
-        assert engine is not None
-        if_match: Optional[str] = None
-        if byte_range is not None:
-            size = byte_range.size
-            bytes_range: Optional[str] = f"bytes={byte_range.offset}-{byte_range.offset + byte_range.size - 1}"
-        else:
-            metadata = self._get_object_metadata(path)
-            size = metadata.content_length
-            bytes_range = None
-            # Bind the GET to the object version the buffer was sized against; if
-            # the object is replaced between the HEAD and the GET the endpoint
-            # returns 412 instead of delivering bytes into a mismatched buffer.
-            if_match = metadata.etag
-
-        def _invoke_api() -> bytearray:
-            buffer = bytearray(size)
-            if size == 0:
-                return buffer
-            get_kwargs: dict[str, Any] = {"Bucket": bucket, "Key": key}
-            if bytes_range is not None:
-                get_kwargs["Range"] = bytes_range
-            if if_match:
-                get_kwargs["IfMatch"] = if_match
-            with engine.transfer(buffer, is_put=False):
-                response = self._s3_client.get_object(**get_kwargs)
-                response["Body"].read()
-                engine.check_reply(response)
-            return buffer
-
-        return self._translate_errors(_invoke_api, operation="GET", bucket=bucket, key=key)
-
-    def _rdma_create_extra(
-        self, bucket: str, attributes: Optional[dict[str, str]], content_type: Optional[str]
-    ) -> dict[str, Any]:
-        extra: dict[str, Any] = {}
-        if content_type:
-            extra["ContentType"] = content_type
-        if self._is_directory_bucket(bucket):
-            extra["StorageClass"] = EXPRESS_ONEZONE_STORAGE_CLASS
-        validated = validate_attributes(attributes)
-        if validated:
-            extra["Metadata"] = validated
-        return extra
-
-    def _rdma_upload(
-        self,
-        remote_path: str,
-        f: Union[str, IO],
-        attributes: Optional[dict[str, str]],
-        content_type: Optional[str],
-    ) -> int:
-        """RDMA upload entry point: single-shot below the part size, multipart above it."""
-        bucket, key = split_path(remote_path)
-        if isinstance(f, str):
-            size = os.path.getsize(f)
-            with open(f, "rb") as fp:
-                if size > self._rdma_multipart_chunksize:
-                    extra = self._rdma_create_extra(bucket, attributes, content_type)
-                    return self._rdma_upload_multipart(bucket, key, fp, size, extra)
-                return self._put_object(remote_path, fp.read(), attributes=attributes, content_type=content_type)
-
-        f.seek(0, io.SEEK_END)
-        size = f.tell()
-        f.seek(0)
-        if size > self._rdma_multipart_chunksize and not isinstance(f, io.StringIO):
-            extra = self._rdma_create_extra(bucket, attributes, content_type)
-            return self._rdma_upload_multipart(bucket, key, f, size, extra)
-        data = f.read()
-        if isinstance(data, str):
-            data = data.encode("utf-8")
-        return self._put_object(remote_path, data, attributes=attributes, content_type=content_type)
-
-    def _rdma_upload_multipart(self, bucket: str, key: str, fp: IO, size: int, extra: dict[str, Any]) -> int:
-        """RDMA multipart upload: each part is transferred as its own registered buffer.
-
-        cuObject transfers one part-sized buffer per ``UploadPart`` (empty HTTP
-        body + RDMA token), each carrying a CRC64NVME the endpoint validates
-        against the RDMA-delivered bytes -- so the 0-byte-save guard applies per
-        part. Multipart is required past the single-PutObject size limit and
-        bounds the size of any one registered buffer / RDMA transfer.
-        """
-        engine = self._rdma_engine
-        assert engine is not None
-        part_size = self._rdma_multipart_chunksize
-
-        def _invoke_api() -> int:
-            upload_id = self._s3_client.create_multipart_upload(Bucket=bucket, Key=key, **extra)["UploadId"]
-            parts: list[dict[str, Any]] = []
-            try:
-                part_number = 1
-                remaining = size
-                while remaining > 0:
-                    n = min(part_size, remaining)
-                    chunk = bytearray()
-                    while len(chunk) < n:
-                        data = fp.read(n - len(chunk))
-                        if not data:
-                            raise RuntimeError(f"unexpected end of input for {bucket}/{key} at part {part_number}")
-                        chunk.extend(data)
-                    checksum = self._rdma_checksum(chunk)
-                    with engine.transfer(chunk, is_put=True):
-                        response = self._s3_client.upload_part(
-                            Bucket=bucket,
-                            Key=key,
-                            UploadId=upload_id,
-                            PartNumber=part_number,
-                            Body=b"",
-                            ChecksumCRC64NVME=checksum,
-                        )
-                        engine.check_reply(response)
-                    parts.append({"PartNumber": part_number, "ETag": response["ETag"]})
-                    remaining -= n
-                    part_number += 1
-                self._s3_client.complete_multipart_upload(
-                    Bucket=bucket, Key=key, UploadId=upload_id, MultipartUpload={"Parts": parts}
-                )
-            except BaseException:
-                try:
-                    self._s3_client.abort_multipart_upload(Bucket=bucket, Key=key, UploadId=upload_id)
-                except Exception:
-                    pass
-                raise
-            return size
-
-        return self._translate_errors(_invoke_api, operation="PUT", bucket=bucket, key=key)
-
     def _put_object(
         self,
         path: str,
@@ -750,9 +535,6 @@ class S3StorageProvider(BaseStorageProvider):
             if validated_attributes:
                 kwargs["Metadata"] = validated_attributes
 
-            if self._rdma_engine is not None:
-                return self._rdma_put(kwargs, body)
-
             # TODO(NGCDP-5804): Add support to update ContentType header in Rust client
             rust_unsupported_feature_keys = {"Metadata", "StorageClass", "IfMatch", "IfNoneMatch", "ContentType"}
             if (
@@ -773,9 +555,6 @@ class S3StorageProvider(BaseStorageProvider):
 
     def _get_object(self, path: str, byte_range: Optional[Range] = None) -> bytes:
         bucket, key = split_path(path)
-
-        if self._rdma_engine is not None:
-            return self._rdma_get(path, bucket, key, byte_range)
 
         def _invoke_api() -> bytes:
             if byte_range:
@@ -1121,9 +900,6 @@ class S3StorageProvider(BaseStorageProvider):
         attributes: Optional[dict[str, str]] = None,
         content_type: Optional[str] = None,
     ) -> int:
-        if self._rdma_engine is not None:
-            return self._rdma_upload(remote_path, f, attributes, content_type)
-
         file_size: int = 0
 
         if isinstance(f, str):
